attempt to restrict purchases
Implemented 'amount bought' tracker and if else statements to check if amountbought is more than produceCost requirement and if so, stop buying. This almost never does anything because in the economy as it is right now the first firm usually has all that is needed. And/Or there are other bugs preventing it from working properly.