fixed bug with buying expand cost :)
so basically, the loop was breaking (not continuing) when it found it wanted to buy 0 of something. this always happenend because the amount to buy in the first place was 0, because the resources given had some with 0 in it. this only affected expand and not upkeep costs calling the function because in expand, if theyre trying to expand it subtracts resources to not buy resources it produces

also implemented limit on purchasing in units of how many items it takes to produce

also nerfed baker back (aw yeah balancing update)

changed order of expadning and upkeep and producing in a firm's tick function

implemented a (necessary) system such that a firm wont produce too much such that it cant expand, if it's trying to expand and the expand requirements are involved in producing and it doesnt have enough for both expand requirement and a produce cycle